RCA/requirements.txt
matplotlib
numpy
olefile
opencv-python
Pillow
protobuf
py-cpuinfo
pycparser
pymongo
pyparsing
PyQt5
PyQt5-sip
PyQtWebEngine
python-dateutil
PyWavelets
scikit-image
scipy
tornado
tqdm
wheel
wrapt
